third_party/WebKit/LayoutTests/http/tests/inspector/network/network-initiator.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>

/*function addClassToDiv()
{
    // We need to setTimeout, to make sure that style recalculation scheduling
    // is really triggered by adding a class to the div.
    window.setTimeout(addClassToDivStep2, 0);
}

function addClassToDivStep2()
{
    document.getElementById("div-without-class").className = "image-background-2";
}*/

function loadData()
{
    var iframe = document.createElement("iframe");
    iframe.src = "resources/network-initiator-frame.html";
    document.body.appendChild(iframe);
}

var test = function()
{
    step2();

    /*function step1()
    {
        InspectorTest.networkManager.addEventListener(WebInspector.NetworkManager.EventTypes.RequestFinished, onRequest);
        InspectorTest.evaluateInPage("addClassToDiv()", step2);
    }

    function onRequest(event)
    {
        var request = event.data;
        if (/size=300/.test(request.url))
            step2();
    }*/

    function step2()
    {
        //if (InspectorTest.findRequestsByURLPattern(/size=300/).length === 0)
        //    return;

        InspectorTest.addConsoleSniffer(step3);
        InspectorTest.evaluateInPage("loadData()");
    }

    function step3()
    {
        function dumpInitiator(url)
        {
            var matching_requests = InspectorTest.findRequestsByURLPattern(new RegExp(url.replace(".", "\\.")));
            if (matching_requests.length === 0) {
                InspectorTest.addResult(url + " NOT FOUND");
                return;
            }
            var request = matching_requests[0];
            var initiator = request.initiator();
            InspectorTest.addResult(request.url + ": " + initiator.type);
            if (initiator.url)
                InspectorTest.addResult("    " + initiator.url + " " + initiator.lineNumber);
            if (initiator.stack) {
                var stackTrace = initiator.stack;
                for (var i = 0; i < stackTrace.callFrames.length; ++i) {
                    var frame = stackTrace.callFrames[i];
                    if (frame.lineNumber) {
                        InspectorTest.addResult("    " + frame.functionName + " " + frame.url + " " + frame.lineNumber);
                        break;
                    }
                }
            }
        }

        dumpInitiator("initiator.css");
        dumpInitiator("size=100");
        //dumpInitiator("size=200");
        //dumpInitiator("size=300");
        dumpInitiator("size=400");
        dumpInitiator("style.css");
        dumpInitiator("empty.html");
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests resources initiator for images initiated by IMG tag, static CSS, CSS class added from JavaScript and XHR.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=65105">Bug 65105</a>
</body>
</html>
