third_party/WebKit/LayoutTests/http/tests/inspector/network/network-initiator-expected.txt

CONSOLE MESSAGE: line 15: Done.
Tests resources initiator for images initiated by IMG tag, static CSS, CSS class added from JavaScript and XHR.

Bug 65105  
http://127.0.0.1:8000/inspector/network/resources/initiator.css: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 3
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=100: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 5
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=400: script
    loadData http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 14
http://127.0.0.1:8000/inspector/network/resources/style.css: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 8
http://127.0.0.1:8000/inspector/network/resources/empty.html: parser
    http://127.0.0.1:8000/inspector/network/resources/network-initiator-frame.html 9
